backend/pyproject.toml
[project]
name = "jobnexus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.13.3",
    "fastapi>=0.128.0",
    "linkedin-scraper>=3.1.0",
    "python-dotenv>=1.2.1",
    "python-jobspy>=1.1.82",
    "uvicorn>=0.40.0",
]
